schedule service updated: update schedules added
