+++ src/services/index.ts
@@ -1,7 +1,5 @@
 export * from "./auth.service";
 
-export * from "./file.service";
-
 export * from "./schedule.service";
 
 export * from "./user.service";
